# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CST Portal Internal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: Account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change temporary/own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /auth/2fa/enable:
    post:
      operationId: enable2fa
      tags: [auth]
      summary: Enable 2FA after verifying TOTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeInput"
      responses:
        "200":
          description: 2FA enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/2fa/verify:
    post:
      operationId: verify2fa
      tags: [auth]
      summary: Verify TOTP code during login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeInput"
      responses:
        "200":
          description: 2FA verified, session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/2fa/disable:
    post:
      operationId: disable2fa
      tags: [auth]
      summary: Disable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorCodeInput"
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        attemptsLeft:
          type: ["integer", "null"]
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, dept, initials, color, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, sector_manager, employee]
        dept:
          type: string
        initials:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [ativo, inativo]
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    TwoFactorCodeInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    ResetPasswordInput:
      type: object
      required: [email, token, newPassword]
      properties:
        email:
          type: string
        token:
          type: string
        newPassword:
          type: string
...