# lib/api-zod/src/generated/types/changePasswordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CST Portal Internal API
 * OpenAPI spec version: 0.1.0
 */

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}