# lib/api-zod/src/generated/types/resetPasswordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CST Portal Internal API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordInput {
  email: string;
  token: string;
  newPassword: string;
}